requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
annotated-types==0.7.0
anyio==4.6.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.18.0
azure-search-documents==11.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.6.1
jiter==0.5.0
msal==1.31.0
msal-extensions==1.2.0
openai==1.51.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
